LTC3: Standardized optoisolator selection.
Replaced the Avago 4N35 on the ignition control schematic with the
Toshiba TLP3542 used on the external power and triggers schematic.
This reduces the number of unique parts used in the design.  The Avago
has an order of magnitude faster switching time than the Toshiba, but
that shouldn't matter for the ignition control circuit; the Toshiba's
2.5A max current is way more than sufficient.